% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_data_addresses}
\alias{calculate_data_addresses}
\title{function to get initial data addresses}
\usage{
calculate_data_addresses()
}
\value{
data frame with info on the initial files
}
\description{
function to get initial data addresses
}
