% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_excel}
\alias{check_for_excel}
\title{Function checking if the uploaded files are Excel files}
\usage{
check_for_excel(files_list)
}
\value{
boolean
}
\description{
Function checking if the uploaded files are Excel files
}
