% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_format}
\alias{check_for_format}
\title{Function checking if the uploaded files are Excel files}
\usage{
check_for_format(files_list)
}
\arguments{
\item{files_list}{files list}
}
\value{
boolean
}
\description{
Function checking if the uploaded files are Excel files
}
