% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading-functions.R
\name{collect_unique_flags}
\alias{collect_unique_flags}
\title{Function collecting unique flags to be used in beat type dropdowns}
\usage{
collect_unique_flags(file_addresses, data_columns, separator)
}
\arguments{
\item{data_columns}{string with numbers of columns with analyzed data}

\item{separator}{separator}

\item{file}{addresses file addresses on disk}
}
\value{
vector
}
\description{
Function collecting unique flags to be used in beat type dropdowns
}
