% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-table-functions.R
\name{correct_dashes}
\alias{correct_dashes}
\title{Function correcting dashes in col names so that it looks nicer}
\usage{
correct_dashes(data_tibble)
}
\arguments{
\item{data_tibble}{tibble with data}
}
\value{
tibble with fixed names
}
\description{
Function correcting dashes in col names so that it looks nicer
}
