% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{cut_incomplete_rows}
\alias{cut_incomplete_rows}
\title{Helper function cutting either the first row of the table, or the last row, or returning the whole table}
\usage{
cut_incomplete_rows(resulting_table, cut_end, return_all)
}
\arguments{
\item{resulting_table}{table with results from a windowing function}

\item{cut_end}{where the cut occurs -at the beginning or at the end or not at all (FALSE, TRUE)}

\item{return_all}{wheter all rows should be returned}
}
\value{
table
}
\description{
Helper function cutting either the first row of the table, or the last row, or returning the whole table
}
