% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-info-module.R
\name{data_info}
\alias{data_info}
\title{module creating the data-info UI and returning its values - handy because it
can be put anywhere (i.e. in modal)}
\usage{
data_info(input, output, session)
}
\arguments{
\item{input}{input shiny variable}

\item{output}{output shiny variable}

\item{session}{input shiny variable}
}
\value{
input reactive values
}
\description{
module creating the data-info UI and returning its values - handy because it
can be put anywhere (i.e. in modal)
}
