% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-upload-and-filter-module.R
\name{data_upload_and_filter}
\alias{data_upload_and_filter}
\title{module with data upload and filtering}
\usage{
data_upload_and_filter(input, output, session)
}
\value{
inputs corresponding to upload and filtering
}
\description{
no input
}
