% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-table-functions.R
\name{description_string}
\alias{description_string}
\title{function returning Poincare descriptors to be shown on the flipside}
\usage{
description_string(data_line, type_of_plot)
}
\arguments{
\item{data_line}{named vector with Poincare descriptors}

\item{type_of_plot}{type of plot}
}
\value{
HTML string
}
\description{
function returning Poincare descriptors to be shown on the flipside
}
