% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading-functions.R
\docType{data}
\name{flags_and_codes}
\alias{flags_and_codes}
\title{List with usual flags coding}
\format{
An object of class \code{list} of length 4.
}
\usage{
flags_and_codes
}
\description{
List with usual flags coding
}
\keyword{datasets}
