% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{get_comparisons_in_windowed_results}
\alias{get_comparisons_in_windowed_results}
\title{Function extracting which comparisons can be applied to a window}
\usage{
get_comparisons_in_windowed_results(windowed_results, comparisons)
}
\arguments{
\item{windowed_results}{windowed results to apply the comparisons}

\item{asym_comparisons}{vector of strings, containing comparisons of the form AR1>DR1, SD1d>SD1a etc for use in dynamic asymmetry}
}
\value{
vector of strings, containing comparisons of the form AR1>DR1, SD1d>SD1a etc for use in dynamic asymmetry
}
\description{
Function extracting which comparisons can be applied to a window
}
