% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-static.R
\name{get_pp_results}
\alias{get_pp_results}
\title{function for getting the results of Poincare Plot analysis}
\usage{
get_pp_results(
  fileAddresses,
  separator = "\\t",
  column_data = c(1, 2),
  minmax = c(0, 3000),
  using_excel = FALSE,
  flags_coding,
  shuffle,
  pnnX_th
)
}
\arguments{
\item{separator}{the separator chosen by the user}

\item{column_data}{a 1x2 vector with the numbers of columns holding RR intervals and annotations}

\item{minmax}{1x2 vector with the maximum and minimum acceptable RR intervals values}

\item{flags_coding}{list with flags_coding}

\item{shuffle}{whether the data should be shuffled}

\item{pnnX_th}{pnnX thresholds vector}

\item{file_addresses}{the addresses of the uploaded file(s)}

\item{using_Excel}{boolean, whether Excel files are used}
}
\value{
the results of Poincare plot analysis
}
\description{
function for getting the results of Poincare Plot analysis
}
