% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-table-functions.R
\name{get_results_matrix}
\alias{get_results_matrix}
\title{function to get results matrix}
\usage{
get_results_matrix(current_values)
}
\arguments{
\item{current_values}{values to be currently displayed}
}
\value{
data_frame with values and View buttons
}
\description{
function to get results matrix
}
