% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{get_single_quality_windowed_results}
\alias{get_single_quality_windowed_results}
\title{Function calculating windowed quality results for a single RR time series}
\usage{
get_single_quality_windowed_results(
  RR,
  time_functions_list = glb_time_functions,
  window_type = "time",
  move_type = "jump",
  window_length = 5,
  cut_end = FALSE,
  return_all = FALSE
)
}
\arguments{
\item{RR}{rr object}

\item{window_type}{type of window, can be time or index}

\item{window_length}{length of the window, in minutes or beats, according to window_type}

\item{slide_type}{how should the window move: slide or jump}
}
\description{
Function calculating windowed quality results for a single RR time series
}
