% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions.R
\name{get_spectral_windowed_results}
\alias{get_spectral_windowed_results}
\title{Function calculating windowed spectral results}
\usage{
get_spectral_windowed_results(
  RR,
  time_functions_list = glb_time_functions,
  window_type = "time",
  move_type = "jump",
  window_length = 5,
  cut_end = FALSE,
  return_all = FALSE
)
}
\arguments{
\item{RR}{rr object}

\item{window_type}{type of window, can be time or index}

\item{window_length}{length of the window, in minutes or beats, according to window_type}

\item{slide_type}{how should the window move: slide or jump}
}
\description{
Function calculating windowed spectral results
}
