% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading-functions.R
\name{getSep}
\alias{getSep}
\title{function to read the separator from input - it is either what has ben selected
or a selection has to be translated into a real separator}
\usage{
getSep(separator)
}
\arguments{
\item{separator}{separator}
}
\value{
separator
}
\description{
function to read the separator from input - it is either what has ben selected
or a selection has to be translated into a real separator
}
