% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\docType{data}
\name{glb_time_functions}
\alias{glb_time_functions}
\title{time window functions as a list}
\format{
An object of class \code{list} of length 4.
}
\usage{
glb_time_functions
}
\description{
time window functions as a list
}
\keyword{datasets}
