% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{alternative for null}
\usage{
x \%||\% y
}
\arguments{
\item{x}{value being checked for null}

\item{y}{default expression value}
}
\value{
x if not null, y otherwise
}
\description{
alternative for null
}
