% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading-module.R
\name{loadData}
\alias{loadData}
\title{data loading module}
\usage{
loadData(input, output, session, dataAddress, separator, data_columns,
  minmax, usingExcel)
}
\arguments{
\item{input}{input shiny variable}

\item{output}{output shiny variable}

\item{session}{input shiny variable}
}
\description{
data loading module
}
