% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loader}
\alias{loader}
\title{Function running loader/spinner}
\usage{
loader(path, timeout = 1000, interval = 1000, sleep = 1000)
}
\arguments{
\item{path}{that to the svg loader}

\item{timeout}{how long to wait to show spinner after registering shiny-busy}

\item{interval}{how check check for shinyBusy}

\item{sleep}{the shortest time spinner shows}
}
\description{
Function running loader/spinner
}
