% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-table-module.R
\name{main_table}
\alias{main_table}
\title{module for displaying main table}
\usage{
main_table(input, output, session, rct_current_pp_values)
}
\arguments{
\item{rct_current_pp_values}{the current values related to poincare plot (to be changed for other types of output)}
}
\value{
DT
}
\description{
module for displaying main table
}
