% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{not_na}
\alias{not_na}
\title{Function for use in Filter equivalent to !is.na}
\usage{
not_na(x)
}
\arguments{
\item{x}{data}
}
\description{
Function for use in Filter equivalent to !is.na
}
