% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-module.R
\name{plots}
\alias{plots}
\title{module to plot various plots related to HRA/HRV}
\usage{
plots(input, output, session, type_of_plot = "poincare",
  rct_data_address, line_number, inp_data_columns, inp_separator,
  inp_minmax, inp_using_excel, inp_variable_name, inp_color)
}
\arguments{
\item{type_of_plot}{plot type (poincare, runs, spectrum etc.)}

\item{rct_data_address}{address of the file to be read}

\item{line_number}{the line holding the name of the file we want to analyze}

\item{inp_data_columns}{numbers of columns holding RR intervals and annotations}

\item{inp_separator}{separator used in file}

\item{inp_minmax}{minimum and maximum values of RR intervals of sinus origin}

\item{inp_using_excel}{whether or not use Excel format}

\item{inp_variable_name}{name of the variable for display}

\item{inp_color}{color for plotting}
}
\value{
plot (also causes side effect of saving the plot to disc)
}
\description{
module to plot various plots related to HRA/HRV
}
