% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{puj}
\alias{puj}
\title{Function formatting p-values - copied from my package shiny-tools}
\usage{
puj(p, rmarkdown = TRUE, digits)
}
\arguments{
\item{p}{p value to be formated}

\item{rmarkdown, }{boolea, whether to use rmarkdown}

\item{digits}{how much to round the p-value}
}
\description{
Function formatting p-values - copied from my package shiny-tools
}
