% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_read_one_file}
\alias{raw_read_one_file}
\title{Function to read one file raw, without any pre or post processing}
\usage{
raw_read_one_file(file_addresses, file_no = 1, separator)
}
\arguments{
\item{file_addresses}{file addresses on disk}

\item{file_no}{number of the file to be read}
}
\value{
data.frame
}
\description{
Function to read one file raw, without any pre or post processing
}
