% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading-module.R
\name{read_and_filter_one_file}
\alias{read_and_filter_one_file}
\title{function to read and filter a single file containing rr intervals and annotations}
\usage{
read_and_filter_one_file(file_addresses, line_number, separator,
  column_data, minmax, using_Excel)
}
\arguments{
\item{file_addresses}{list with name, size, type and path of files with rr intervals}

\item{line_number}{the position of the file being loaded on the list}

\item{separator}{separator for the loaded file (tab, comma etc.)}

\item{column_data}{text, two numbers, the numbers of the RR column and the annotation column (typically "1 2")}

\item{minmax}{tex, two numbers, the minimum RR to be counted as RR of sinus origin, likewise for maximum}

\item{using_Excel}{boolean, whether the files are Excel files}
}
\value{
list with two elements, RR intervals column and annotations column
}
\description{
function to read and filter a single file containing rr intervals and annotations
}
