% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-loading.R
\name{read_numbers_from_field}
\alias{read_numbers_from_field}
\title{function to read numbers from input windows}
\usage{
read_numbers_from_field(list_argument)
}
\arguments{
\item{list_argument}{string holding the numbers keyed into the input field}
}
\value{
numeric vector holding the values
}
\description{
function to read numbers from input windows
}
