% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{round_and_summarize_dynamic_asym}
\alias{round_and_summarize_dynamic_asym}
\title{Function adding dynamic asymmetry tests and rounding values}
\usage{
round_and_summarize_dynamic_asym(
  window,
  round_digits = 3,
  p_digits = 4,
  asym_comparisons = NULL
)
}
\arguments{
\item{window}{result of a numerical function applied to an RR window}

\item{round_digits}{how much to round the descirptors}

\item{p_digits}{how should the p-value be rounded}

\item{asym_comparisons}{vector of strings, containing comparisons of the form AR1>DR1, SD1d>SD1a etc for use in dynamic asymmetry}
}
\description{
Function adding dynamic asymmetry tests and rounding values
}
