% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_table}
\alias{sample_table}
\title{Function producing sample table for preview}
\usage{
sample_table(samp_table)
}
\arguments{
\item{samp_table}{data frame}
}
\value{
data frame
}
\description{
Function producing sample table for preview
}
