% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{shuffle_in_windows}
\alias{shuffle_in_windows}
\title{Function shuffling data in windows if necessary}
\usage{
shuffle_in_windows(window_table, shuffle, rr_index)
}
\arguments{
\item{window_table}{table with windowed signal (single window)}

\item{shuffle}{whether the table should be shuffled (shuffled if value == "window")}

\item{rr_index}{which column contains RR intervals}
}
\description{
Function shuffling data in windows if necessary
}
