% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-results-module.R
\name{single_results}
\alias{single_results}
\title{function to build the table with results for}
\usage{
single_results(input, output, session, type_of_plot, line_number,
  rct_current_pp_values)
}
\description{
function to build the table with results for
}
