% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-results-module.R
\name{single_resultsUI}
\alias{single_resultsUI}
\title{module to show results for a single recording on the flip side}
\usage{
single_resultsUI(id)
}
\description{
module to show results for a single recording on the flip side
}
