% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{sort_ardrn}
\alias{sort_ardrn}
\title{Sorting function for runs}
\usage{
sort_ardrn(ardrn_vec, type = "DR")
}
\arguments{
\item{ardrn_vec}{vector of runs names}

\item{type}{DR, AR or N}
}
\description{
Sorting function for runs
}
