% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{sort_out_NAs}
\alias{sort_out_NAs}
\title{Function checking wheter a NA should be replaced with 0 or not}
\usage{
sort_out_NAs(results_table)
}
\arguments{
\item{results_table}{table of results which has NAs - they will be replaced with 0s if the whole line contains NAs}
}
\description{
Function checking wheter a NA should be replaced with 0 or not
}
