% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{sort_out_runs}
\alias{sort_out_runs}
\title{Function to sort out the coluns of the runs results table}
\usage{
sort_out_runs(results)
}
\arguments{
\item{results}{windowed results}
}
\description{
Function to sort out the coluns of the runs results table
}
