% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{sort_ps}
\alias{sort_ps}
\title{Sorting function for comparisons}
\usage{
sort_ps(ardr_p, stub = "_prop")
}
\arguments{
\item{ardr_p}{comparison names}

\item{stub}{what the stub of the comparison is (_prop or _pVal)}
}
\description{
Sorting function for comparisons
}
