% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-crunching-functions-dynamic.R
\name{get_dynamic_spectral_results}
\alias{get_dynamic_spectral_results}
\title{function for getting the results of Poincare Plot analysis}
\usage{
get_dynamic_spectral_results(
  fileAddresses,
  use_ULF = "No",
  time_functions_list = glb_time_functions,
  separator = "\\t",
  column_data = c(1, 2),
  minmax = c(0, 3000),
  using_excel = FALSE,
  window_type,
  time_unit,
  move_type,
  window_length,
  clicked_file,
  flags_coding,
  shuffle,
  tolerance
)
}
\arguments{
\item{separator}{the separator chosen by the user}

\item{column_data}{a 1x2 vector with the numbers of columns holding RR intervals and annotations}

\item{minmax}{1x2 vector with the maximum and minimum acceptable RR intervals values}

\item{window_type}{string, jumping or sliding}

\item{time_unit}{unit of time (minutes or seconds)}

\item{move_type}{string, time based or index based}

\item{window_length}{numeric, window length}

\item{clicked_file}{number of clicked file or NULL}

\item{flags_coding}{list with flags_coding}

\item{shuffle}{whether the data should be shuffled}

\item{tolerance}{what is the maximum data loss in a single window in dynamic analysis that should be tolerated}

\item{file_addresses}{the addresses of the uploaded file(s)}

\item{using_Excel}{boolean, whether Excel files are used}
}
\value{
the results of Poincare plot analysis
}
\description{
function for getting the results of Poincare Plot analysis
}
