% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-table-functions.R
\name{poincare_description_string}
\alias{poincare_description_string}
\title{function returning Poincare descriptors to be shown on the flipside}
\usage{
poincare_description_string(data_line)
}
\arguments{
\item{data_line}{named vector with Poincare descriptors}
}
\value{
HTML string
}
\description{
function returning Poincare descriptors to be shown on the flipside
}
